Find Longest Substring WithoutRepeating Characters
You are given a string S, and your task is to find the length of the longest substring that contains no repeating characters. A substring is a contiguous block of characters in the string.
In this problem, you need to find the length of the longest substring where all the characters are unique. The substring can be formed by starting at any position in the string, but it must not contain any repeated characters.

Input:
A string S, where 1≤∣S∣≤105(length of string).

Output:
An integer representing the length of the longest substring without repeating characters.

Examples:
Example 1
Input: S = "abcabcbb"
Output: 3
Explanation: The answer is "abc", with a length of 3. Even though the string contains repeated characters, the longest substring without duplicates is "abc".

Constraints:
1≤∣S∣≤105
The string contains only printable ASCII characters.

Test Cases:
Input: S = "abcabcbb"
Output: 3
Input: S = "bbbbb"
Output: 1
Input: S = "pwwkew"
Output: 3
Input: S = "abcdefgh"
Output: 8
Input: S = "a"
Output: 1

Edge Cases:
Single character string: The longest substring without repeating characters is the character itself.
All characters are unique: If no characters repeat, the entire string is the longest substring.
All characters are the same: If all characters are identical, the longest substring will be of length 1.