Check if a Binary Tree is Symmetric
You are given the root of a binary tree. Your task is to determine whether the tree is symmetric. A binary tree is symmetric if the left and right subtrees are mirror images of each other.

Input:
The root of the binary tree.

Output:
Return true if the binary tree is symmetric, otherwise return false.

Examples:
Example 1
Input: [1, 2, 2, 3, 4, 4, 3]
Output: true

This binary tree is symmetric as the left and right subtrees are mirror images of each other.

Constraints:
The number of nodes in the tree is between 1 and 104
The value of each node is between −100 and 100.

Test Cases:
Input: [1, 2, 2, 3, 4, 4, 3]
Output: false
Input: [1, 2, 2, null, 3, null, 3]
Output: false
Input: [1]
Output: true
Input: []
Output: true
Input: [1, 2, 2, 3, null, null, 3]
Output: false

Edge Cases:
An empty tree is symmetric by definition. Return true.
A single-node tree is symmetric since there are no children to compare. Return true.
If there is one child but not the other, the tree is not symmetric. Return false.